refactor(py): fold Apply Filter button test into existing e2e suite
Moved the filter-button regression test from its own repro file into
Test01HelloApp, where it belongs alongside the other hello-app end-to-end
scenarios.
